+++ config/mongoose.js
@@ -1,0 +1,17 @@
+const mongoose = require("mongoose");
+if (process.env.NODE_ENV !== "production") {
+  require("dotenv").config();
+}
+mongoose.connect(process.env.MONGODB_URI, {
+  useNewUrlParser: true,
+  useUnifiedTopology: true,
+});
+// 取得資料庫連線狀態
+const db = mongoose.connection;
+db.on("error", () => {
+  console.log("mongodb error!");
+});
+db.once("open", () => {
+  console.log("mongodb connected!");
+});
+module.exports = db;
