+++ package.json
@@ -15,6 +15,7 @@
     "dotenv": "^16.0.3",
     "express": "^4.18.2",
     "express-handlebars": "^3.0.0",
+    "method-override": "^3.0.0",
     "mongoose": "^5.9.7"
   }
 }
