+++ views/edit.handlebars
@@ -1,6 +1,6 @@
 <h1 class="mb-1 restaurant-show-title">編輯餐廳 - {{restaurantData.name}}</h1>
 <div class="container">
-  <form action="/restaurants/{{restaurantData._id}}?_method=PUT" method="POST">
+  <form action="./restaurants/{{restaurantData._id}}?_method=PUT" method="POST">
     <div class="mb-3">
       <label for="name" class="form-label">店名<span
           class="text-danger"
