+++ views/index.handlebars
@@ -65,7 +65,10 @@
               <a href="/restaurants/{{this._id}}"><button
                   class="btn btn-link btn-sm"
                 >Details</button></a>
-              <form action="./restaurants/{{this._id}}/delete" method="POST">
+              <form
+                action="./restaurants/{{this._id}}?_method=DELETE"
+                method="POST"
+              >
                 <button
                   type="submit"
                   class="btn btn-danger btn-sm text-white"
